OTel.API.Common.Attributes

Documentation

data Attrs (af :: AttrsFor)

emptyAttrs :: Attrs af

nullAttrs :: Attrs af -> Bool

sizeAttrs :: Attrs af -> Int

memberAttrs :: Key a -> Attrs af -> Bool

lookupAttrs :: forall a af. KnownAttrType a => Key a -> Attrs af -> Maybe (Attr a)

foldMapWithKeyAttrs :: forall m af. Monoid m => (forall a. Key a -> Attr a -> m) -> Attrs af -> m

filterWithKeyAttrs :: forall af. (forall a. Key a -> Attr a -> Bool) -> Attrs af -> Attrs af

mapWithKeyAttrs :: forall af. (forall a. Key a -> Attr a -> Attr a) -> Attrs af -> Attrs af

convertWithKeyAttrs :: forall af. (forall a. Key a -> Attr a -> SomeAttr) -> Attrs af -> Attrs af

Equivalent to mapWithKeyAttrs but allows for changing both the type and value of each attribute rather than just the value of the attribute.

droppedAttrsCount :: Attrs af -> Int

class WithAttrs (a :: Type) where

Associated Types

type WithAttrsAttrType a :: AttrsFor

Methods

(.:@) :: a -> AttrsBuilder (WithAttrsAttrType a) -> a infixr 6

data AttrsBuilder (af :: AttrsFor)

jsonAttrs :: forall a af. ToJSON a => Text -> a -> AttrsBuilder af

data AttrsFor

Constructors

AttrsForResource

AttrsForSpan

AttrsForSpanEvent

AttrsForSpanLink

data AttrsLimits (af :: AttrsFor)

defaultAttrsLimits :: AttrsLimits af

attrsLimitsCount :: AttrsLimits af -> Maybe Int

attrsLimitsValueLength :: AttrsLimits af -> Maybe Int

data SomeAttr where

Constructors

SomeAttr :: Attr a -> SomeAttr

data Attr a

Constructors

Attr

Fields

attrType :: AttrType a

attrVal :: a

asTextAttr :: Attr a -> Attr Text

Convert an attribute to a text attribute.

This function is identity if the attribute already is a text attribute. Otherwise, the attribute value is passed to show.

data AttrVals a

data AttrType (a :: Type) where

Constructors

AttrTypeText :: AttrType Text

AttrTypeBool :: AttrType Bool

AttrTypeDouble :: AttrType Double

AttrTypeInt :: AttrType Int64

AttrTypeTextArray :: AttrType (AttrVals Text)

AttrTypeBoolArray :: AttrType (AttrVals Bool)

AttrTypeDoubleArray :: AttrType (AttrVals Double)

AttrTypeIntArray :: AttrType (AttrVals Int64)

class KnownAttrType a where

Methods

attrTypeVal :: Proxy a -> AttrType a

class KnownAttrType to => ToAttrVal from to | from -> to where

Methods

toAttrVal :: from -> to