OTel.API.Common.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

General

class KV (kv :: Type) where

Associated Types

type KVConstraints kv :: Type -> Type -> Constraint

Methods

(.@) :: KVConstraints kv from to => Key to -> from -> kv infixr 8

class (k ~ Text, v ~ Text) => IsTextKV k v

newtype Key a

Constructors

Key

Fields

unKey :: Text

newtype Timestamp

Constructors

Timestamp

Fields

unTimestamp :: Integer

nanoseconds

timestampFromNanoseconds :: Integer -> Timestamp

timestampToNanoseconds :: Timestamp -> Integer

data TimestampSource

Constructors

TimestampSourceNow

TimestampSourceAt Timestamp

data InstrumentationScope

Constructors

InstrumentationScope

Fields

instrumentationScopeName :: InstrumentationScopeName

instrumentationScopeVersion :: Maybe Version

instrumentationScopeSchemaURL :: Maybe SchemaURL

defaultInstrumentationScope :: InstrumentationScope

newtype InstrumentationScopeName

Constructors

InstrumentationScopeName

Fields

unInstrumentationScopeName :: Text

newtype Version

Constructors

Version

Fields

unVersion :: Text

newtype SchemaURL

Constructors

SchemaURL

Fields

unSchemaURL :: Text

schemaURLFromText :: Text -> Either Text SchemaURL

schemaURLToText :: SchemaURL -> Text

Attributes

class WithAttrs (a :: Type) where

Associated Types

type WithAttrsAttrType a :: AttrsFor

Methods

(.:@) :: a -> AttrsBuilder (WithAttrsAttrType a) -> a infixr 6

data Attrs (af :: AttrsFor)

Constructors

Attrs

Fields

attrsMap :: HashMap Text SomeAttr

attrsDropped :: Int

emptyAttrs :: Attrs af

nullAttrs :: Attrs af -> Bool

sizeAttrs :: Attrs af -> Int

memberAttrs :: Key a -> Attrs af -> Bool

lookupAttrs :: forall a af. KnownAttrType a => Key a -> Attrs af -> Maybe (Attr a)

foldMapWithKeyAttrs :: forall m af. Monoid m => (forall a. Key a -> Attr a -> m) -> Attrs af -> m

filterWithKeyAttrs :: forall af. (forall a. Key a -> Attr a -> Bool) -> Attrs af -> Attrs af

mapWithKeyAttrs :: forall af. (forall a. Key a -> Attr a -> Attr a) -> Attrs af -> Attrs af

convertWithKeyAttrs :: forall af. (forall a. Key a -> Attr a -> SomeAttr) -> Attrs af -> Attrs af

Equivalent to mapWithKeyAttrs but allows for changing both the type and value of each attribute rather than just the value of the attribute.

droppedAttrsCount :: Attrs af -> Int

newtype AttrsBuilder (af :: AttrsFor)

Constructors

AttrsBuilder

Fields

unAttrsBuilder :: Int -> DList AttrsBuilderElem

runAttrsBuilder :: AttrsBuilder af -> AttrsLimits af -> Attrs af

jsonAttrs :: forall a af. ToJSON a => Text -> a -> AttrsBuilder af

data AttrsAcc

Constructors

AttrsAcc

Fields

attrsAccMap :: HashMap Text SomeAttr

attrsAccMapSize :: Int

attrsAccDropped :: Int

data AttrsBuilderElem

Constructors

AttrsBuilderElem

Fields

attrsBuilderElemKey :: Text

attrsBuilderElemVal :: SomeAttr

data AttrsFor

Constructors

AttrsForResource

AttrsForSpan

AttrsForSpanEvent

AttrsForSpanLink

data AttrsLimits (af :: AttrsFor)

Constructors

AttrsLimits

Fields

attrsLimitsCount :: Maybe Int

attrsLimitsValueLength :: Maybe Int

defaultAttrsLimits :: AttrsLimits af

data SomeAttr where

Constructors

SomeAttr :: Attr a -> SomeAttr

data Attr a

Constructors

Attr

Fields

attrType :: AttrType a

attrVal :: a

asTextAttr :: Attr a -> Attr Text

Convert an attribute to a text attribute.

This function is identity if the attribute already is a text attribute. Otherwise, the attribute value is passed to show.

newtype AttrVals a

Constructors

AttrVals

Fields

unAttrVals :: Vector a

data AttrType (a :: Type) where

Constructors

AttrTypeText :: AttrType Text

AttrTypeBool :: AttrType Bool

AttrTypeDouble :: AttrType Double

AttrTypeInt :: AttrType Int64

AttrTypeTextArray :: AttrType (AttrVals Text)

AttrTypeBoolArray :: AttrType (AttrVals Bool)

AttrTypeDoubleArray :: AttrType (AttrVals Double)

AttrTypeIntArray :: AttrType (AttrVals Int64)

class KnownAttrType a where

Methods

attrTypeVal :: Proxy a -> AttrType a

class KnownAttrType to => ToAttrVal from to | from -> to where

Methods

toAttrVal :: from -> to

with :: a -> (a -> b) -> b

newtype OnException a

Constructors

OnException

Fields

runOnException :: SomeException -> [SeriesElem] -> LoggingT IO a

askException :: OnException SomeException

askExceptionMetadata :: OnException [SeriesElem]

newtype OnTimeout a

Constructors

OnTimeout

Fields

runOnTimeout :: Int -> [SeriesElem] -> LoggingT IO a

askTimeoutMicros :: OnTimeout Int

askTimeoutMetadata :: OnTimeout [SeriesElem]

data BufferedLoggerSpec

Constructors

BufferedLoggerSpec

Fields

bufferedLoggerSpecShouldBuffer :: Loc -> LogSource -> LogLevel -> LogStr -> Bool

Predicate that determines if a log message should be buffered. The default is no buffering, so all log messages are logged immediately.

bufferedLoggerSpecLogger :: Logger

The logger to wrap. For all unbuffered log messages, these are passed to bufferedLoggerSpecLogger immediately. For buffered log messages, aggregates will eventually be passed to bufferedLoggerSpecLogger on the configured period (see bufferedLoggerSpecFlushPeriod).

bufferedLoggerSpecFlushPeriod :: Int

Buffered logs are regularly flushed from internal storage on this period (in microseconds). The default is 5 minutes.

bufferedLoggerSpecFlushTimeout :: Int

Max amount of time allowed for flushing buffered logs (in microseconds). The default is 10 seconds.

bufferedLoggerSpecOnFlushTimeout :: BufferedLogs -> OnTimeout ()

Handler that is run if flushing buffered logs takes longer than bufferedLoggerSpecFlushTimeout. The default is to log the timeout using bufferedLoggerSpecOnFlushExceptionLogger. This may be overridden if the timeout needs to be reported or dealt with via some means other than logging. Note that if the handler throws a synchronous exception, that exception will be caught and ignored.

The input argument is all remaining buffered logs that were unable to be flushed within the timeout.

bufferedLoggerSpecOnFlushException :: BufferedLog -> BufferedLogAgg -> OnException ()

Handler that is run if a synchronous exception is encountered during log flushing. The default is to log the exception using bufferedLoggerSpecOnFlushExceptionLogger. This may be overridden if the exception needs to be reported or dealt with via some means other than logging. Note that if the handler rethrows the exception, the exception will be caught and ignored.

The input arguments are the specific buffered log and aggregate that encountered an exception during flushing.

bufferedLoggerSpecOnFlushExceptionLogger :: Logger

The bufferedLoggerSpecOnFlushTimeout and bufferedLoggerSpecOnFlushException handlers are run using this logger. This logger is intentionally different from bufferedLoggerSpecLogger, as bufferedLoggerSpecLogger may have ultimately been what triggered running either handler in the first place.

bufferedLoggerSpecIncludeLogAggregate :: BufferedLog -> BufferedLogAgg -> [SeriesElem] -> LogStr

The internal storage tracks a count of each buffered log message's occurrences (rather than tracking each occurrence) and each occurrence's metadata (convenience for monad-logger-aeson users). This function enables including this aggregate info in the log message when it is eventually logged. The default is to not include the aggregate info.

monad-logger-aeson users may find it convenient to set this function to includeLogAggregateViaAeson Just. This will include the aggregate in full. If there is a risk that a buffered log's aggregated metadata may exceed a logging system's max payload per message, Just can be replaced with a function that more carefully summarizes the metadata. For example, to ignore the metadata altogether and instead only log the count of aggregated messages, the following could be used:

includeLogAggregateViaAeson \bufferedLogAgg ->
  Just $ object ["count" .= bufferedLogAggCount bufferedLogAgg]

defaultBufferedLoggerSpec :: BufferedLoggerSpec

includeLogAggregateViaAeson

Arguments

:: forall a. ToJSON a

=> (BufferedLogAgg -> Maybe a)
Summarizes the aggregated info

-> BufferedLog

-> BufferedLogAgg

-> [SeriesElem]

-> LogStr

withBufferedLogger :: forall m a. MonadUnliftIO m => BufferedLoggerSpec -> (Logger -> m a) -> m a

withBufferedLoggerIO :: forall a. BufferedLoggerSpec -> (Logger -> IO a) -> IO a

type BufferedLogs = HashMap BufferedLog BufferedLogAgg

insertBufferedLog :: BufferedLog -> KeyMap Value -> BufferedLogs -> BufferedLogs

insertBufferedLogWithAgg :: BufferedLog -> BufferedLogAgg -> BufferedLogs -> BufferedLogs

data BufferedLog

Constructors

BufferedLog

Fields

bufferedLogLoc :: Loc

bufferedLogLogSource :: LogSource

bufferedLogLogLevel :: LogLevel

bufferedLogLogStr :: Either ByteString Text

Right when the log message was parsed as a Message, Left otherwise.

toBufferedLog :: Loc -> LogSource -> LogLevel -> LogStr -> (BufferedLog, KeyMap Value)

data BufferedLogAgg

Constructors

BufferedLogAgg

Fields

bufferedLogAggCount :: Int

bufferedLogAggMetas :: DList (KeyMap Value)

type Logger = Loc -> LogSource -> LogLevel -> LogStr -> IO ()